Add Update and Remove Works and Financial

+++ src/services/RemovePhilanthropyService.ts
@@ -10,6 +10,10 @@ class RemovePhilanthropyService {
       throw new Error("Philanthropy not found");
     }
 
+    if (philanthropyToRemove.type !== "philanthropy") {
+      throw new Error("Philanthropy type incorrect");
+    }
+
     await calendarRepository.delete({ id: id });
 
     return philanthropyToRemove;
